feat: implement initial database migration and model setup for team planning (#15)
- Added initial migration to create database schema for Boards, Features, Sprints, TeamMembers, UserStories, and TeamMemberSprints.
- Updated model snapshot to reflect new database structure.
- Modified TeamMemberSprint model to change CapacityDev and CapacityTest types from double to int.
- Implemented methods in TeamRepository and ITeamRepository for deleting team members.
- Enhanced TeamService to support adding, updating, and deleting team members with capacity management.
- Updated API endpoints and service methods for team member management.
- Improved frontend components for team member actions, including edit and delete functionalities.
- Added validation for team member name and capacity inputs in the UI.

+++ backend/pi-planning-backend/DTOs/BoardResponseDto.cs
@@ -45,7 +45,7 @@ namespace PiPlanningBackend.DTOs
     public class TeamMemberSprintDto
     {
         public int SprintId { get; set; }
-        public double CapacityDev { get; set; }
-        public double CapacityTest { get; set; }
+        public int CapacityDev { get; set; }
+        public int CapacityTest { get; set; }
     }
 }
